Risk Management Simulation Project
NIST risk assessment - Created a NIST risk assessment of all identified risk register items and suggested NIST controls to mitigate the identified risks, bringing the risk down to an acceptable level based on the Risk Matrix. 

Conducted a full risk assessment on the identified risk items included in the Risk Register for the Zombie Healthcare System organization

Once all the items were evaluated, identified and mitigated the risk to an acceptable level, the information was summarized for the ZHS leadership team.

The report to the leadership team is a presentation using slides which included:
Risk Register with a summary of the risks.
NIST Controls explaining the controls used and their effectiveness.